debug.
git-svn-id: svn+ssh://svn.pricetx.org/home/svn/numberguesser@18 d9013ba4-f9c6-e311-acb0-001cc0fd6b5d

+++ NumberGuesser.c
@@ -76,7 +76,7 @@ main()
 	
 	/* Initialise random number generator */
 	srand((unsigned int)(time(NULL)));
-	printf("THE *TOTALLY RANDOM* TIME IS %d", time(NULL));
+	printf("THE *TOTALLY RANDOM* TIME IS %d", (unsigned int)time(NULL));
 
         /* run gamemode based on argument */
 	switch (mode) {
